// src/app/shared/Models/IBrand.ts
export interface IBrand {
    title: string
    description: string | null
    isActive: boolean | null
    summary: string | null
    products: any | null
    created: string | null
    createdBY: any | null
    lastModified: any | null
    lastModifiedBy: any | null
    id: number
    isDeleted: boolean | null
  }